+++ projects/go-lib/src/lib/components/go-copy/go-copy.module.ts
@@ -3,12 +3,10 @@ import { CommonModule } from '@angular/common';
 
 import { GoCopyComponent } from './go-copy.component';
 import { GoIconModule } from '../go-icon/go-icon.module';
-import { GoCopyDocLinkDirective } from './go-copy-doc-link.directive';
 
 @NgModule({
   declarations: [
     GoCopyComponent,
-    GoCopyDocLinkDirective
   ],
   imports: [
     CommonModule,
@@ -16,7 +14,6 @@ import { GoCopyDocLinkDirective } from './go-copy-doc-link.directive';
   ],
   exports: [
     GoCopyComponent,
-    GoCopyDocLinkDirective
   ]
 })
